starting with the skeleton
